Add partial logic for fetch operation
The fetch operation to get the wineList from the database is more complicated than originally expected. Doing so requires that the user auth token is passed to the database which requires creating a context for the user auth token to be able to use the signed in state in multiple components through the site.